requirements.txt
djangorestframework==3.14.0
drf-spectacular==0.26.5
